+++ systems/unconverted/generic/client/character.vue
@@ -177,7 +177,7 @@
     </div>
 </template>
 
-<style lang="sass" src="./character.scss"></style>
+<style lang="sass" src="character.scss"></style>
 
 <script type="text/babel">
     import { modal } from 'vueboot';
@@ -189,7 +189,7 @@
     import counter from './components/counter.vue';
     import roll from './components/roll.vue';
     import statblock from './components/statblock.vue';
-    import notes from '../../../client/components/notes/notes.vue';
+    import notes from '../../../../client/components/notes/notes.vue';
     import AddEditStatModal from './modals/statAddEdit.vue';
 
     export default {
